Use Discord media proxy urls whenever possible (#12)
* Use Discord media proxy urls whenever possible

This consists of using an image's proxy_url instead of url when it is
available. (It will not be available when the image is not hosted on
Discord.)
The Discord media proxy (media.discordapp.net) is a better way to
access images because it is CORS enabled, so it does not require going
through the pxls CORS proxy to access it. It also caches images, so
that they can still be accessed a bit after being deleted.
We also switch to using the media proxy when fetching emoji images.

* Remove _EmbedMediaProxy type annotation

I cannot find a satisfying type for this argument since multiple types
of objects can have proxy_url and proxy, and the types are not even
always specified properly in disnake.
Instead just remove it.