Merge downstream into origin to prepare for repository transfer (#16)
* Add final canvases for images 62 to 74

* Create Dockerfile and docker-compose.yaml for
containerization

* Remove unused pxls_validate cookie

* Remove lyrics functionality

* Change command permission decorator for >prefix

New: Bot owner or anyone with the administrator permission

* Add .dockerignore file to project

* Refactor emoji IDs 🔄

Replace outdated and inconsistent emoji IDs with updated ones for
improved visual presentation and clarity throughout the codebase.
No functional changes are introduced.

* feat: Add restart command for owners

This commit adds a new command, "restart", to the Utility cog. This
command allows bot owners to restart the bot by calling the command. It
sends a message indicating that the bot is restarting and then calls the
"exit()" function to gracefully terminate the bot process.

No issue referenced.

* Refactor DbStatsManager SQL queries to filter
records by canvas code

This change modifies the SQL queries in the DbStatsManager class to
include the filtering of record data based on the canvas code.
Previously, the queries were only filtering records based on datetime.
However, with this change, records are now filtered based on both
datetime and canvas code.